+++ CHANGELOG.md
@@ -26,6 +26,8 @@
     - Use `Criteo#loadBid(AdUnit, BidResponseListener)` method to provide `Bid` object
   - Make `CriteoNativeAdListener` an interface instead of an abstract class
   - Replace `View` parameter by a `CriteoBannerView` in `CriteoBannerAdListener#onAdReceived`
+  - Remove `CriteoBannerAdListener#onAdClosed`
+  - Remove `CriteoBannerAdListener#onAdOpened`
 
 - Features
   - CriteoBannerView is now a custom view that can be included directly in a layout file.